OCT-111: moved GetCatalogsByOwnerUsernameQuery
